# sysclient/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SysInspect API",
    "description": "SysInspect Web API for interacting with the master interface.",
    "license": {
      "name": "",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.1"
  },
  "paths": {
    "/api/v1/authenticate": {
      "post": {
        "tags": [
          "System"
        ],
        "description": "Authenticates a user using configured authentication method. The payload must be a base64-encoded, RSA-encrypted JSON object with username and password fields as follows:\n\n```json\n{\n\"username\": \"darth_vader\",\n\"password\": \"I am your father\",\n\"pubkey\": \"...\"\n}\n```\n\nIf the API is in development mode, it will return a static token without actual authentication.",
        "operationId": "authenticateUser",
        "requestBody": {
          "description": "Base64-encoded, RSA-encrypted JSON containing username and password. See description for details.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Authentication successful. Returns a session ID (sid) if credentials are valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                },
                "example": {
                  "status": "authenticated",
                  "sid": "session-id"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request. Returned if payload is missing, invalid, or credentials are incorrect.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                },
                "example": {
                  "status": "error",
                  "sid": null,
                  "error": "Invalid payload"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthRequest": {
        "type": "object",
        "required": [
          "payload",
          "pubkey"
        ],
        "properties": {
          "payload": {
            "type": "string",
            "description": "Base64-encoded, RSA-encrypted JSON: {\"username\": \"...\", \"password\": \"...\", \"pubkey\": \"...\"}"
          },
          "pubkey": {
            "type": "string"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "required": [
          "status",
          "sid_cipher",
          "symkey_cipher",
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "sid_cipher": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "symkey_cipher": {
            "type": "string"
          }
        }
      },
...
    }
  }
}